# openapi.yaml
openapi: 3.0.3
info:
  title: DurianPay Dashboard API
  version: "1.0.0"
servers:
  - url: http://localhost:8080

components:
  parameters:
    sort:
      name: sort
      in: query
      description: >
        Comma-separated sort fields. Prefix `-` for descending, no prefix for ascending.
        Example: `-created_at` (desc), `amount` (asc)
      required: false
      schema:
        type: string
        example: "-created_at"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "unauthorized"
        message:
          type: string
          example: "missing or invalid token"

    User:
      type: object
      properties:
        email:
          type: string
          example: "cs@test.com"
        role:
          type: string
          example: "cs"
        token:
          type: string
          example: "eyJhbGci..."

    Payment:
      type: object
      properties:
        id:
          type: string
          example: "1"
        merchant_name:
          type: string
          example: "Tokopedia"
        amount:
          type: string
          example: "150000"
        status:
          type: string
          enum: [completed, processing, failed]
          example: "completed"
        created_at:
          type: string
          format: date-time

  responses:
    LoginResponse:
      description: Successful login, returns JWT token and user info
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'

    PaymentListResponse:
      description: List of payments
      content:
        application/json:
          schema:
            type: object
            properties:
              payments:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'

    UnauthorizedError:
      description: Authentication failed or missing credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthenticated:
              summary: Missing or invalid token
              value:
                code: "unauthorized"
                message: "missing or invalid token"

paths:
  /dashboard/v1/auth/login:
    post:
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: "cs@test.com"
                password:
                  type: string
                  example: "password"
      responses:
        "200":
          $ref: '#/components/responses/LoginResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'

  /dashboard/v1/payments:
    get:
      summary: List payments, optionally filtered by status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/sort'
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [completed, processing, failed]
          description: Filter payments by status
      responses:
        "200":
          $ref: '#/components/responses/PaymentListResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'